backend/requirements.txt
pandas
requests
openai
python-dotenv
scikit-learn
Flask
Flask-Cors